split of file parsing and cleanup file_import
